Cleaned up code involving extraNonce1, identified problem with cgminer 3.7.2
In cgminer 3.7.2 (cgminer.c:6058) "memcpy(pool->coinbase +
pool->nonce2_offset, &pool->nonce2, sizeof(uint32_t))"

This line is flawed because the size of nonce2 is not always uint32_t,
in the most common proxypool usage, it is set to 2. Thus when it copys
2 extra bytes, a part of the coinbase is over written with the higher
two bytes of pool->nonce2. This does not cause a problem (luckly) with
P2Pool since the coinbase2 is always set to "00000000", thus only if the
miner is faster than 256 tera herz will it start to cause a problem.

However, other pool software does not use pure zeros for coinbase2, thus
causing all submitted shares to be invalid on cgminer 3.7.2