Fix space leak caused by Chan
Sometimes the connection is not fast enough to catch up to the work
update, this results the Chan slowly filling up with useless work that
still needs to be sent, slowly ballooning memory.

We change it so that each client only holds the serialised version of
the current work update, and let the work update thread signal a push to
the client.